Update NP_NAT defenition to fix TypeError bug

+++ trading_calendars/calendar_helpers.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 NANOSECONDS_PER_MINUTE = int(6e10)
 
-NP_NAT = np.array([pd.NaT], dtype=np.int64)[0]
+NP_NAT = np.array([pd.NaT])[0]
 
 
 def next_divider_idx(dividers, minute_val):
